Refactor: move utils out of `UtExecutionInstrumentation.kt`
