+++ utbot-framework/src/main/kotlin/org/utbot/framework/concrete/UtExecutionInstrumentation.kt
@@ -4,7 +4,6 @@ import java.security.ProtectionDomain
 import java.util.IdentityHashMap
 import kotlin.reflect.jvm.javaMethod
 import org.utbot.framework.UtSettings
-import org.utbot.framework.assemble.AssembleModelGenerator
 import org.utbot.framework.concrete.constructors.ConstructOnlyUserClassesOrCachedObjectsStrategy
 import org.utbot.framework.concrete.constructors.UtModelConstructor
 import org.utbot.framework.concrete.mock.InstrumentationContext
@@ -13,9 +12,7 @@ import org.utbot.framework.concrete.phases.start
 import org.utbot.framework.plugin.api.Coverage
 import org.utbot.framework.plugin.api.EnvironmentModels
 import org.utbot.framework.plugin.api.FieldId
-import org.utbot.framework.plugin.api.UtAssembleModel
 import org.utbot.framework.plugin.api.UtExecutionResult
-import org.utbot.framework.plugin.api.UtExecutionSuccess
 import org.utbot.framework.plugin.api.UtInstrumentation
 import org.utbot.framework.plugin.api.UtModel
 import org.utbot.framework.plugin.api.util.singleExecutableId
@@ -45,53 +42,12 @@ class UtConcreteExecutionResult(
     val result: UtExecutionResult,
     val coverage: Coverage
 ) {
-
     override fun toString(): String = buildString {
         appendLine("UtConcreteExecutionResult(")
         appendLine("stateAfter=$stateAfter")
         appendLine("result=$result")
         appendLine("coverage=$coverage)")
     }
-}
-
-private fun UtConcreteExecutionResult.updateWithAssembleModels(
-    assembledUtModels: IdentityHashMap<UtModel, UtModel>
-): UtConcreteExecutionResult {
-    val toAssemble: (UtModel) -> UtModel = { assembledUtModels.getOrDefault(it, it) }
-
-    val resolvedStateAfter = EnvironmentModels(
-        stateAfter.thisInstance?.let { toAssemble(it) },
-        stateAfter.parameters.map { toAssemble(it) },
-        stateAfter.statics.mapValues { toAssemble(it.value) }
-    )
-    val resolvedResult =
-        (result as? UtExecutionSuccess)?.model?.let { UtExecutionSuccess(toAssemble(it)) } ?: result
-
-    return UtConcreteExecutionResult(
-        resolvedStateAfter,
-        resolvedResult,
-        coverage
-    )
-}
-
-/**
- * Tries to convert all models from [UtExecutionResult] to [UtAssembleModel] if possible.
- *
- * @return [UtConcreteExecutionResult] with converted models.
- */
-fun UtConcreteExecutionResult.convertToAssemble(packageName: String): UtConcreteExecutionResult {
-    val allModels = collectAllModels()
-
-    val modelsToAssembleModels = AssembleModelGenerator(packageName).createAssembleModels(allModels)
-    return updateWithAssembleModels(modelsToAssembleModels)
-}
-
-private fun UtConcreteExecutionResult.collectAllModels(): List<UtModel> {
-    val allModels = listOfNotNull(stateAfter.thisInstance).toMutableList()
-    allModels += stateAfter.parameters
-    allModels += stateAfter.statics.values
-    allModels += listOfNotNull((result as? UtExecutionSuccess)?.model)
-    return allModels
 }
 
 object UtExecutionInstrumentation : Instrumentation<UtConcreteExecutionResult> {
